Handle cannotClose option for caseTable and caseCard.
